Updates to prepare obsidianbot for 100% online time.
Added stop, shutdown, and updates to have the bot disconnect from voice
when a playlist ends and when it is the only one connected in the voice
channel. These conditions are both checked when a song ends. Further,
the launcher gui has been updated and several bugs have been fixed.