Require Java 11, update gradle & deps.
Oracle's JDK is going to be used less and less in favor of OpenJDK. JavaFX is not available in the base OpenJDK, therefore we need to migrate to OpenJFX. This is only available for Java 9+. Ideally the project would be moved straight to the latest version, Java 13, however Gradle does not yet support it. As such, the project will be moved to Java 11 LTS until support for Java 13 is expanded.

OpenJFX is now bundled with AventiBot using the OpenJFX gradle plugin. It should work with any OpenJDK distribution which is version 11 or higher. Only tested with the hotspot vm.